pyproject.toml
[project]
name = "drission-shopping-mcp"
version = "0.1.0"
description = "Naver Shopping + DrissionPage based MCP server for product search and detail extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.20.0",
  "httpx>=0.27.0",
  "DrissionPage>=4.1.0,<5",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.3.0",
  "starlette>=0.46.0",
  "uvicorn>=0.34.0",
  "pydantic>=2.11.0",
  "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.11.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["shopping_mcp", "shopping_mcp.adapters"]

[tool.ruff]
line-length = 100
